+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 ### Added
-- Added webApplicationLink and webApplicationLinkRepository constructor params
+- Added webApplication constructor parameter
 
 ## 1.0.0 - 2023-09-07